Moved HighchartsComponent to a seperate folder.
