+++ tools/gulptasks/dashboards/api-docs.json
@@ -28,9 +28,11 @@
         "../../../ts/Dashboards/Layout/Row.ts",
         "../../../ts/Dashboards/PluginHandler.ts",
         "../../../ts/Dashboards/Plugins/DataGridComponent.ts",
-        "../../../ts/Dashboards/Plugins/HighchartsComponent.ts",
+        "../../../ts/Dashboards/Plugins/HighchartsComponent/HighchartsComponent.ts",
+        "../../../ts/Dashboards/Plugins/HighchartsComponent/HighchartsComponentOptions.ts",
         "../../../ts/Dashboards/Plugins/KPIComponent.ts",
-        "../../../ts/Dashboards/Plugins/NavigatorComponentOptions.ts",
+        "../../../ts/Dashboards/Plugins/NavigatorComponent/NavigatorComponent.ts",
+        "../../../ts/Dashboards/Plugins/NavigatorComponent/NavigatorComponentOptions.ts",
         "../../../ts/Data/Connectors/CSVConnectorOptions.d.ts",
         "../../../ts/Data/Connectors/JSONConnectorOptions.d.ts",
         "../../../ts/Data/Connectors/DataConnectorOptions.d.ts",
