+++ ts/Dashboards/Plugins/HighchartsPlugin.ts
@@ -22,8 +22,8 @@
 import type PluginHandler from '../PluginHandler';
 import type { Highcharts as H } from './HighchartsTypes';
 
-import HighchartsComponent from './HighchartsComponent.js';
-import HighchartsSyncHandlers from './HighchartsSyncHandlers.js';
+import HighchartsComponent from './HighchartsComponent/HighchartsComponent.js';
+import HighchartsSyncHandlers from './HighchartsComponent/HighchartsSyncHandlers.js';
 import KPIComponent from './KPIComponent.js';
 import NavigatorComponent from './NavigatorComponent/NavigatorComponent.js';
